feat: engine infrastructure — ImGui REPL console, SDL init, HiDPI, gamepad docs

+++ content/render/tilemap2_impl.cc
@@ -593,8 +593,8 @@ URGE_DEFINE_OVERRIDE_ATTRIBUTE(
     },
     {
       DISPOSE_CHECK;
-      if (repeat_.x != value) {
-        repeat_.x = value;
+      if (repeat_.x != static_cast<int32_t>(value)) {
+        repeat_.x = static_cast<int32_t>(value);
         map_buffer_dirty_ = true;
       }
     });
@@ -609,8 +609,8 @@ URGE_DEFINE_OVERRIDE_ATTRIBUTE(
     },
     {
       DISPOSE_CHECK;
-      if (repeat_.y != value) {
-        repeat_.y = value;
+      if (repeat_.y != static_cast<int32_t>(value)) {
+        repeat_.y = static_cast<int32_t>(value);
         map_buffer_dirty_ = true;
       }
     });
